Add predefined portage variables
When running through portage, 50+ predefined variables are created
and/or consumed by the build system outside of the ebuild script.  Some
of them are common words that might be likely to be used by non-portage
shell scripts, so we don't want to simply declare them as internal shell
variables.

Instead, parameterize the functions that return internal lists based on
the Parameters in use and only include portage variables when
isPortageBuild is true.