Problem: then word in pcs scripts confuse ShellCheck
Pacemaker-related scripts often use 'then' word unquoted and it does
work in bash. Example:

```
pcs constraint order VirtualIP then dummy_resource kind=Optional
```

This instruction causes SC1010 warning which seems to be a
false-positive.

It seems like 'then' word itself is not a keyword in bash. I guess that
it is understood as a keyword only in context of `if` clause.

Solution:
Exclude 'then' from the list of terminal tokens in `readNormalWord`
function.